requirements.txt
# CHS-SDK Examples Dependencies

# Core dependencies
chs-core>=0.1.0
chs-agent>=0.1.0
chs-ai>=0.1.0
chs-analyze>=0.1.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web framework (for web demo)
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Utilities
pyyaml>=5.4.0
click>=8.0.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0